pyproject.toml
[build-system]
requires = ["setuptools>=69.5.1"]
build-backend = "setuptools.build_meta"

# to install the whole project into site-packages (where your interpreter is)
#   then run: pip install -e .
#   (HIGHLY recommended within a conda env or virtual env)
[project]
name = "realtime-isaac-item-detector"
version = "1.0.0"
description = "Computer vision-powered Binding of Isaac Item Detector"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE.txt"}
keywords = ["Computer Vision", "Binding of Isaac", "Object Detection", "pytorch"]

authors = [
    {name = "Hamza Khan"}
]

maintainers = [
    {name = "Hamza Khan"}
]

# to install everything run: pip install -e .
dependencies = [
    "requests",
    "beautifulsoup4",
    "opencv-python",
    "pillow",
    "matplotlib",
    "torch",
    "torchvision",
    "torchaudio",
    "ultralytics",
    "numpy",
    "ipykernel",
    "mss",
    "pywin32",
    "pygetwindow",
    "flask",
    "flask-socketio"
]

[project.urls]
"Homepage" = "https://github.com/khan0617/realtime-binding-of-isaac-item-detector"

# Development dependencies
# pip install -e .[lint,test]
# pip install -e .[dev]
[project.optional-dependencies]
lint = [
    # code style formatting, run like `black .`
    "black",

    # automatic sorting of imports, run like `isort .`
    "isort",

    # more static code linting
    "pylint"
]

test = [
    # test execution platform
    "pytest",

    # provides test coverage % from test runs
    "pytest-cov",

    # pytest wrapper for python's unittest.mock
    "pytest-mock"
]

dev = [
    # python type checking, run using `mypy .` to check everything.
    "mypy",

    # types for the beautifulsoup library for mypy
    "types-beautifulsoup4",

    # types for requests library for mypy
    "types-requests",

    # types for yaml
    "types-PyYAML",

    # types for tqdm
    "types-tqdm",

    # actions for pre commit hooks to run formatters, linters etc.
    # to configure the pre-commit hooks, run "pre-commit install"
    # this will add these hooks to the local .git file.
    "pre-commit"
]


# pip install -e .[docs]
# to generate docs, we first need to generate all the .rst files for our .py modules:
#   run `sphinx-autodoc -o source src` from the project root
#   This creates the *.rst in source/
# next, run `make html`, this creates build/ in the project root.
# to update the sphinx docs on github pages, move the content in build/html to docs/
docs = [
    "sphinx",
    "sphinx_rtd_theme",  # Optional, for the ReadTheDocs theme
    "sphinx-autodoc-typehints",  # Optional, for type hints in docstrings
]


[tool.setuptools.packages.find]
where = ["src"]
exclude = ["test", "test.*"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = true
src_paths = ["src/"]

[tool.pylint]
max-line-length = 120
min-similarity-lines = 4
disable = [
    "missing-module-docstring",
    "missing-function-docstring"
]
output-format = "colorized"
ignore = [
    "test"
]

# see here for mypy pyproject.toml https://mypy.readthedocs.io/en/stable/config_file.html#using-a-pyproject-toml-file
[tool.mypy]
mypy_path = "./src"
follow_imports = "silent"
python_version = "3.11"
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = false

# Configuration for pytest
# https://docs.pytest.org/en/latest/reference/customize.html#pyproject-toml
[tool.pytest.ini_options]
testpaths = ["test"]
norecursedirs = [
    ".vscode",
    "__pycache__"
]